Made the time code more sane and changed some types and their meaning around.
Time is now measured in milliseconds, instead of the old format which used the upper 32bit for the seconds the lower 32bit for the milliseconds.
There is now also a type for holding unix timestamps (in seconds) as well as convenient functions to convert the two time types around.

There is also code from the FreeBSD project which is used for 64bit integer division (see /sys/libc/bsd/README.md)